+++ python/examples/wavelet_buffer_example.py
@@ -7,6 +7,8 @@ from wavelet_buffer import (
     energy_distribution,
 )
 
+from wavelet_buffer.img import HslJpeg
+
 # Processing 1-D time series
 signal = np.array([1, 2, 3, 3, 2, 1, 0, -1, -2, -3, -3, -2, -1, 0, 1, 2, 3, 3, 2, 1])
 
@@ -40,3 +42,26 @@ buffer = WaveletBuffer(
     decomposition_steps=4,
     wavelet_type=WaveletType.DB1,
 )
+
+with open("tests/fixtures/pandas.jpg", "rb") as f:
+    blob = f.read()
+
+code = HslJpeg()
+img = code.decode(blob)
+print(f"Image shape {img.shape}")
+
+buffer.decompose(img, denoiser)
+
+print(f"Channels {len(buffer.decompositions)}")
+print(
+    f"Subbands for channel=0: decomposition_steps*3 +1  {len(buffer.decompositions[0])}"
+)
+print(f"The last subband is approximation:  {buffer.decompositions[0][12]}")
+
+print(f"Reconstructed image x2 smaller and save")
+smaller_img: np.ndarray = buffer.compose(scale_factor=1)
+print(smaller_img.shape)
+blob = code.encode(smaller_img)
+
+with open("2x_pandas.jpg", "wb") as f:
+    f.write(blob)