+++ python/src/denoise.cc
@@ -1,31 +1,23 @@
-// Copyright 2020 PANDA GmbH
+// Copyright 2020-2022 PANDA GmbH
+
+#include "denoise.h"
+
 #include <wavelet_buffer/denoise_algorithms.h>
 
-#include <pybind11/pybind11.h>
-#include <pybind11/operators.h>
-
-namespace py = pybind11;
-
 void WrapDenoiseAlgorithms(py::module *m) {
-  using drift::NullDenoiseAlgorithm;
-  using drift::SimpleDenoiseAlgorithm;
-  using drift::ThresholdAbsDenoiseAlgorithm;
-  using drift::DenoiseAlgorithm;
-
   auto module = m->def_submodule("denoise");
 
-  auto base = py::class_<DenoiseAlgorithm<float>>(module, "Base");
-  py::class_<NullDenoiseAlgorithm<float>, DenoiseAlgorithm<float>>
-      (module, "Null")
+  auto base = py::class_<drift::DenoiseAlgorithm<float>>(module, "Base");
+
+  py::class_<drift::NullDenoiseAlgorithm<float>,
+             drift::DenoiseAlgorithm<float>>(module, "Null")
       .def(py::init());
 
-  py::class_<SimpleDenoiseAlgorithm<float>, DenoiseAlgorithm<float>>
-      (module, "Simple")
-      .def(py::init<float>(),
-           py::arg("compression_level"));
+  py::class_<drift::SimpleDenoiseAlgorithm<float>,
+             drift::DenoiseAlgorithm<float>>(module, "Simple")
+      .def(py::init<float>(), py::arg("compression_level"));
 
-  py::class_<ThresholdAbsDenoiseAlgorithm<float>, DenoiseAlgorithm<float>>
-      (module, "Threshold")
-      .def(py::init<float, float>(),
-          py::arg("a"), py::arg("b"));
+  py::class_<drift::ThresholdAbsDenoiseAlgorithm<float>,
+             drift::DenoiseAlgorithm<float>>(module, "Threshold")
+      .def(py::init<float, float>(), py::arg("a"), py::arg("b"));
 }