+++ python/src/wavelet_buffer_view.cc
@@ -1,20 +1,15 @@
-// Copyright 2021 PANDA GmbH
+// Copyright 2021-2022 PANDA GmbH
+
+#include "wavelet_buffer_view.h"
 
 #include <wavelet_buffer/wavelet_buffer_view.h>
-#include <pybind11/numpy.h>
-#include <pybind11/pybind11.h>
 
 #include "blaze_utils.h"
+#include "wavelet_buffer_view_proxy.h"
 
 namespace py = pybind11;
 
-using drift::WaveletBuffer;
 using Denoiser = drift::DenoiseAlgorithm<float>;
-using drift::Signal1D;
-using drift::Signal2D;
-using drift::SignalN2D;
-
-#include "wavelet_buffer_view_proxy.h"
 
 /**
  *
@@ -24,7 +19,7 @@ using drift::SignalN2D;
  */
 void DecomposeN2DSignal(WaveletBufferViewProxy *self, const py::array &data,
                         const Denoiser &denoiser) {
-  auto &buffer = self->buffer.cast<WaveletBuffer &>();
+  auto &buffer = self->buffer.cast<drift::WaveletBuffer &>();
   const auto &params = buffer.parameters();
 
   if (data.ndim() != 3) {
@@ -58,8 +53,8 @@ void DecomposeN2DSignal(WaveletBufferViewProxy *self, const py::array &data,
 
 py::array_t<float> ComposeN2DSignal(WaveletBufferViewProxy *self,
                                     int scale_factor) {
-  SignalN2D data;
-  auto &buffer = self->buffer.cast<WaveletBuffer &>();
+  drift::SignalN2D data;
+  auto &buffer = self->buffer.cast<drift::WaveletBuffer &>();
   if (!buffer(self->start_signal, self->count).Compose(&data, scale_factor)) {
     throw py::buffer_error("Failed to compose data");
   }
@@ -73,7 +68,8 @@ void WrapWaveletViewBuffer(py::module *m) {
       "decompose",
       [](WaveletBufferViewProxy &self, const py::array &data,
          const Denoiser &denoiser) {
-        const auto &params = self.buffer.cast<WaveletBuffer &>().parameters();
+        const auto &params =
+            self.buffer.cast<drift::WaveletBuffer &>().parameters();
         if (params.dimension() != 2) {
           throw py::buffer_error("Unsupported number of dimensions");
         }
@@ -84,7 +80,8 @@ void WrapWaveletViewBuffer(py::module *m) {
   cls.def(
       "compose",
       [](WaveletBufferViewProxy &self, int scale_factor) {
-        const auto &params = self.buffer.cast<WaveletBuffer &>().parameters();
+        const auto &params =
+            self.buffer.cast<drift::WaveletBuffer &>().parameters();
         if (params.dimension() != 2) {
           throw py::buffer_error("Unsupported number of dimensions");
         }